Burt atrix

The Burt matrix is a quadratic matrix where each row and column corresponds to a category in one of the variables. The entries of the matrix are the frequencies of the corresponding combination of categories.

Burt(x)

Arguments

x
A dataframe with factor variables or a contingency table.

Value

A matrix.

See also

Examples

require(MASS)

#> Loading required package: MASS

Burt(housing)

#>                Sat:Low Sat:Medium Sat:High Infl:Low Infl:Medium Infl:High
#> Sat:Low            567          0        0      282         206        79
#> Sat:Medium           0        446        0      170         189        87
#> Sat:High             0          0      668      175         264       229
#> Infl:Low           282        170      175      627           0         0
#> Infl:Medium        206        189      264        0         659         0
#> Infl:High           79         87      229        0           0       395
#> Type:Tower          99        101      200      140         172        88
#> Type:Apartment     271        192      302      268         297       200
#> Type:Atrium         64         79       96       95          84        60
#> Type:Terrace       133         74       70      124         106        47
#> Cont:Low           262        178      273      234         279       200
#> Cont:High          305        268      395      393         380       195
#>                Type:Tower Type:Apartment Type:Atrium Type:Terrace Cont:Low
#> Sat:Low                99            271          64          133      262
#> Sat:Medium            101            192          79           74      178
#> Sat:High              200            302          96           70      273
#> Infl:Low              140            268          95          124      234
#> Infl:Medium           172            297          84          106      279
#> Infl:High              88            200          60           47      200
#> Type:Tower            400              0           0            0      219
#> Type:Apartment          0            765           0            0      317
#> Type:Atrium             0              0         239            0       82
#> Type:Terrace            0              0           0          277       95
#> Cont:Low              219            317          82           95      713
#> Cont:High             181            448         157          182        0
#>                Cont:High
#> Sat:Low              305
#> Sat:Medium           268
#> Sat:High             395
#> Infl:Low             393
#> Infl:Medium          380
#> Infl:High            195
#> Type:Tower           181
#> Type:Apartment       448
#> Type:Atrium          157
#> Type:Terrace         182
#> Cont:Low               0
#> Cont:High            968

th <- xtabs(Freq~Sat+Infl+Type, data = housing)
Burt(th)

#>                Sat:Low Sat:Medium Sat:High Infl:Low Infl:Medium Infl:High
#> Sat:Low            567          0        0      282         206        79
#> Sat:Medium           0        446        0      170         189        87
#> Sat:High             0          0      668      175         264       229
#> Infl:Low           282        170      175      627           0         0
#> Infl:Medium        206        189      264        0         659         0
#> Infl:High           79         87      229        0           0       395
#> Type:Tower          99        101      200      140         172        88
#> Type:Apartment     271        192      302      268         297       200
#> Type:Atrium         64         79       96       95          84        60
#> Type:Terrace       133         74       70      124         106        47
#>                Type:Tower Type:Apartment Type:Atrium Type:Terrace
#> Sat:Low                99            271          64          133
#> Sat:Medium            101            192          79           74
#> Sat:High              200            302          96           70
#> Infl:Low              140            268          95          124
#> Infl:Medium           172            297          84          106
#> Infl:High              88            200          60           47
#> Type:Tower            400              0           0            0
#> Type:Apartment          0            765           0            0
#> Type:Atrium             0              0         239            0
#> Type:Terrace            0              0           0          277