The car customers dataset from 1983

This dataset is taken from the website of the Department of Statistics, University of Munich.
The data are based upon a poll from a german car-company. In 1983 questionnaires were sent to 2000 customers, who had purchased a new car approximately three months earlier. The point of interest was the degree of satisfaction, reasons for the particular choice, consumer profile, etc. Participation was of course voluntary. Only 1182 persons answered the questions and after removing forms with "missing values" only 793 questionnaires remained. Each form contained 46 questions, which resulted in a dataset of 46 covariates with 793 observations each. Due to the abundance of ordinal and categorical covariates the dataset is particularly suited for generalized linear models.

data(carcustomers)

Format

A data frame with 774 observations on the following 47 variables.

model

a factor with levels A B C D

gear

a factor with levels 4-gear 5-gear (overdrive) 5-gear (sport) Automatic

lease

a factor with levels bought leased

usage

a factor with levels business private private and business

premod

a factor with levels Audi BMW 3er BMW 5er BMW 7er Ford Mercedes Benz Opel other origin Volkswagen

other

a factor with levels No Yes, both Yes, other manufact Yes, same manufact.

testdrv

influence on buying decision: testdrive

promotion

influence on buying decision: promotion

exp

influence on buying decision: experience

recom

influence on buying decision: recommendation

clear

influence on buying decision: clearness

eco

influence on buying decision: economical aspects

drvchar

influence on buying decision: driving character

service

influence on buying decision: service

interior

influence on buying decision: interior

quality

influence on buying decision: overall quality

tech

influence on buying decision: technical aspects

evo

influence on buying decision: evolution

comfort

influence on buying decision: comfort

reliab

influence on buying decision: reliability

handling

influence on buying decision: handling

prestige

influence on buying decision: prestige

concept

influence on buying decision: overall concept

char

influence on buying decision: character

power

influence on buying decision: engine power

valdecr

influence on buying decision:value decrease

styling

influence on buying decision: styling

safety

influence on buying decision:safety

sport

influence on buying decision: sportive

fcons

influence on buying decision: fuel consumption

space

influence on buying decision: available space

sat

overall satisfaction with the car: 1(very satisfied) to 5(not satisfied)

adv1

satisfaction with concept and styling: a factor with levels does not suit neither nor suits

adv2

satisfaction with body/bare essentials: a factor with levels does not suit neither nor suits

adv3

satisfaction with chassis/drive/gearshift: a factor with levels does not suit neither nor suits

adv4

satisfaction with engine/power: a factor with levels does not suit neither nor suits

adv5

satisfaction with electronics: a factor with levels does not suit neither nor suits

adv6

satisfaction with financial aspects: a factor with levels does not suit neither nor suits

adv7

asatisfaction with equipment: a factor with levels does not suit neither nor suits

spoco

balance variables: a factor with levels comfort could be better handling could be better well balanced

faver

usual driving style: a factor with levels economical extreme normal powerful

sspeed

usual speed (Autobahn): a factor with levels >110 mph 60-80 mph 81-9g mph 96-110 mph

sfcons

satisfaction with fuel consumption: a factor with levels Appropriate Definitely too high Just okay Pleasingly low

sex

customer's gender: a factor with levels Female Male

prof

customer's profession: a factor with levels Employee/Workman Free lanced Self employed

family

customers's family type: a factor with levels >3 persons 1-2 persons

Freq

the weighting variable

Source

http://www.stat.uni-muenchen.de/service/datenarchiv/auto/auto_e.html

Examples

data(Autos)

#> Warning: data set ‘Autos’ not found

## maybe str(Autos) ; plot(Autos) ...