Fluctuation diagram with additional dendrograms

Plots a fluctuation diagram via fluctile and adds dendrograms for the rows and columns to it.

tfluctile(x, tree = NULL, dims = c(1, 2), tw = 0.2, border = NULL, 
 shape = "r", dir = "b", just = "c",
tile.col = hsv(0.1, 0.1, 0.1, alpha = 0.6), bg.col = "lightgrey", vp = NULL, 
lab.opt = list(), …)

Arguments

x
The two-way table or matrix with the data.

tree
A list with tree objects. This may be NULL or will be disregarded if x has an attribute attr(x, "tree") which should also be a list. The latter way is the standard for objects returned by optile.list or optile.

dims
If x has more than two dimensions this vector of length 2 indicates which variables to plot.

tw
The proportion of the total space to the left and at the top which is used for the dendrogram.

border
How much space is left white around the dendrogram.

shape
Instead of rectangles ("r") it is possible to use circles ("c"), diamonds ("d") or octagons ("o"). The arguments dir and just work for rectangular shapes only.

dir
The bar/rectangle direction: "v" and "h" stand for vertical or horizontal bars. "b" stands for "both" and leads to standard fluctuation diagrams with quadratic rectangles. Use "n" for a same-binsize-plot

just
A shortcut version of the argument used in grid for the anchorpoint of the rectangles: "rb" is equivalent to c("right", "bottom"), "t" is equivalent to "ct" or c("centre", "top") and so on. See examles.

tile.col
The tile color.

bg.col
A background color for the cells.

vp
A viewport to which the plot should be added or NULL.

lab.opt
A list with options for the labels. Currently lab.cex and abbrev work. Also lwd and line.col are the line width and the line color for the dendrogram.

…
further args

Value

invisible(TRUE)

See also

Examples

# NOT RUN {
library(amap)
hc1 <- hcluster(t(plants[,-1]), method="manhattan", link = "ward")
hc2 <- hcluster(t(plants[,-1]), method="manhattan", link = "complete")

hclist <- list(hc1, hc2)
tfluctile( tt<-optile(hclist, k= c(8,8) ) )

s1 <- subtree(hc1, k = 12)
s2 <- subtree(hc2, k = 10)

tfluctile( table(s1$data, s2$data), tree = list(s1,s2))
# }